+++ web/src/api/schema.d.ts
@@ -142,8 +142,6 @@ export interface components {
   pathItems: never;
 }
 
-export type $defs = Record<string, never>;
-
 export type external = Record<string, never>;
 
 export interface operations {