+++ rust/stackable-cockpit/src/platform/credentials.rs
@@ -4,7 +4,7 @@ use kube::{core::DynamicObject, ResourceExt};
 use serde::Serialize;
 use snafu::{OptionExt, ResultExt, Snafu};
 
-use crate::utils::k8s;
+use crate::utils::k8s::{self, Client};
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
 
@@ -34,7 +34,7 @@ impl Display for Credentials {
 /// and/or `password_key` are not found or the product does not provide
 /// any credentials.
 pub async fn get(
-    kube_client: &k8s::Client,
+    client: &Client,
     product_name: &str,
     stacklet: &DynamicObject,
 ) -> Result<Option<Credentials>> {
@@ -56,7 +56,7 @@ pub async fn get(
                 .as_str()
                 .context(NoSecretSnafu)?;
 
-            kube_client
+            client
                 .get_credentials_from_secret(
                     secret_name,
                     &stacklet.namespace().unwrap(),
@@ -71,7 +71,7 @@ pub async fn get(
                 .as_str()
                 .context(NoSecretSnafu)?;
 
-            kube_client
+            client
                 .get_credentials_from_secret(
                     secret_name,
                     &stacklet.namespace().unwrap(),