+++ rust/stackable-cockpit/src/platform/stacklet/mod.rs
@@ -12,7 +12,7 @@ use crate::{
     constants::PRODUCT_NAMES,
     platform::{credentials, service},
     utils::{
-        k8s::{self, ConditionsExt},
+        k8s::{self, Client, ConditionsExt},
         string::Casing,
     },
 };
@@ -65,27 +65,27 @@ pub enum Error {
 /// namespaces are returned. If `namespace` is [`Some`], only stacklets installed
 /// in the specified namespace are returned. The `options` allow further
 /// customization of the returned information.
-pub async fn list_stacklets(namespace: Option<&str>) -> Result<Vec<Stacklet>, Error> {
-    let kube_client = k8s::Client::new().await.context(KubeClientCreateSnafu)?;
-
-    let mut stacklets = list_stackable_stacklets(&kube_client, namespace).await?;
-    stacklets.extend(grafana::list(&kube_client, namespace).await?);
-    stacklets.extend(minio::list(&kube_client, namespace).await?);
-    stacklets.extend(opensearch::list(&kube_client, namespace).await?);
-    stacklets.extend(prometheus::list(&kube_client, namespace).await?);
+pub async fn list_stacklets(
+    client: &Client,
+    namespace: Option<&str>,
+) -> Result<Vec<Stacklet>, Error> {
+    let mut stacklets = list_stackable_stacklets(client, namespace).await?;
+    stacklets.extend(grafana::list(client, namespace).await?);
+    stacklets.extend(minio::list(client, namespace).await?);
+    stacklets.extend(opensearch::list(client, namespace).await?);
+    stacklets.extend(prometheus::list(client, namespace).await?);
 
     Ok(stacklets)
 }
 
 pub async fn get_credentials_for_product(
+    client: &Client,
     namespace: &str,
     object_name: &str,
     product_name: &str,
 ) -> Result<Option<credentials::Credentials>, Error> {
-    let kube_client = k8s::Client::new().await.context(KubeClientCreateSnafu)?;
-
     let product_gvk = gvk_from_product_name(product_name);
-    let product_cluster = match kube_client
+    let product_cluster = match client
         .get_namespaced_object(namespace, object_name, &product_gvk)
         .await
         .context(KubeClientFetchSnafu)?
@@ -99,7 +99,7 @@ pub async fn get_credentials_for_product(
         }
     };
 
-    let credentials = match credentials::get(&kube_client, product_name, &product_cluster).await {
+    let credentials = match credentials::get(client, product_name, &product_cluster).await {
         Ok(credentials) => credentials,
         Err(credentials::Error::NoSecret) => None,
         Err(credentials::Error::KubeClientFetch { source }) => {
@@ -111,14 +111,14 @@ pub async fn get_credentials_for_product(
 }
 
 async fn list_stackable_stacklets(
-    kube_client: &k8s::Client,
+    client: &Client,
     namespace: Option<&str>,
 ) -> Result<Vec<Stacklet>, Error> {
     let product_list = build_products_gvk_list(PRODUCT_NAMES);
     let mut stacklets = Vec::new();
 
     for (product_name, product_gvk) in product_list {
-        let objects = match kube_client
+        let objects = match client
             .list_objects(&product_gvk, namespace)
             .await
             .context(KubeClientFetchSnafu)?
@@ -148,7 +148,7 @@ async fn list_stackable_stacklets(
             };
 
             let endpoints =
-                service::get_endpoints(kube_client, product_name, &object_name, &object_namespace)
+                service::get_endpoints(client, product_name, &object_name, &object_namespace)
                     .await
                     .context(ServiceFetchSnafu)?;