+++ rust/stackable-cockpitd/src/handlers/stacklets.rs
@@ -1,5 +1,5 @@
 use axum::{routing::get, Json, Router};
-use stackable_cockpit::platform;
+use stackable_cockpit::{platform, utils::k8s::Client};
 
 pub use stackable_cockpit::platform::stacklet::Stacklet;
 
@@ -9,9 +9,16 @@ pub fn router() -> Router {
 }
 
 /// Retrieves all stacklets.
+// TODO: Add proper error handling
 #[utoipa::path(get, path = "/stacklets", responses(
     (status = 200, body = Vec<Stacklet>),
 ))]
 pub async fn get_stacklets() -> Json<Vec<Stacklet>> {
-    Json(platform::stacklet::list_stacklets(None).await.unwrap())
+    let client = Client::new().await.expect("failed to construct k8s client");
+
+    Json(
+        platform::stacklet::list_stacklets(&client, None)
+            .await
+            .expect("failed to list stacklets"),
+    )
 }