+++ rust/stackablectl/CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- chore: Avoid unnecessary `k8s::Client` creations ([#295]).
+
 ## [24.3.3] - 2024-05-13
 
 - Bump Rust, Go and Node dependencies ([#238]).
 
 [#238]: https://github.com/stackabletech/stackable-cockpit/pull/238
+[#295]: https://github.com/stackabletech/stackable-cockpit/pull/295
 
 ## [24.3.2] - 2024-04-25