+++ rust/stackablectl/CHANGELOG.md
@@ -6,13 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
-- Re-run GVK discovery after resolution failure ([#XXX]).
+- Re-run GVK discovery after resolution failure ([#294]).
 
 ## [24.3.3] - 2024-05-13
 
 - Bump Rust, Go and Node dependencies ([#238]).
 
 [#238]: https://github.com/stackabletech/stackable-cockpit/pull/238
+[#294]: https://github.com/stackabletech/stackable-cockpit/pull/294
 
 ## [24.3.2] - 2024-04-25