+++ rust/stackablectl/CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Re-run GVK discovery after resolution failure ([#294]).
+
 ## [24.3.3] - 2024-05-13
 
 - Bump Rust, Go and Node dependencies ([#238]).
 
 [#238]: https://github.com/stackabletech/stackable-cockpit/pull/238
+[#294]: https://github.com/stackabletech/stackable-cockpit/pull/294
 
 ## [24.3.2] - 2024-04-25