+++ rust/stackablectl/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Avoid unnecessary `k8s::Client` creations ([#295]).
 - Re-run GVK discovery after resolution failure ([#294]).
 
 ## [24.3.3] - 2024-05-13
@@ -14,6 +15,7 @@ All notable changes to this project will be documented in this file.
 
 [#238]: https://github.com/stackabletech/stackable-cockpit/pull/238
 [#294]: https://github.com/stackabletech/stackable-cockpit/pull/294
+[#295]: https://github.com/stackabletech/stackable-cockpit/pull/295
 
 ## [24.3.2] - 2024-04-25